fix(cleanup): Minor cleanup items * Add name map generate to the run.sh script * Add APT_API_TOKEN env to docker-compose
