+++ container/docker-compose.yml
@@ -1,6 +1,6 @@
 version: "3.8"
 services:
-  vaa:
+  snack:
     environment:
       - CHOKIDAR_USEPOLLING
       - DOC_APP_PATH
@@ -9,3 +9,4 @@ services:
       - GIT_APP_URL
       - KEG_DOCKER_EXEC
       - NODE_ENV
+      - APT_API_TOKEN
